Fix Intness get all tests working again

+++ src/bcs/BoundaryCondition.jl
@@ -1,11 +1,11 @@
 """
     A BoundaryCondition is a Kernel... it just computes differently
 
-    A BoundaryCondition _must_ supply a bid::Array{Int32} variable
+    A BoundaryCondition _must_ supply a bid::Array{Int64} variable
     that specifies which boundary IDs this BC is applied to
 """
 abstract BoundaryCondition <: Kernel
-  # bids::Array{Int32}
+  # bids::Array{Int64}
 
 " A NodalBC gets applied _at_ nodes and overwrites residual / jacobian entries "
 abstract NodalBC <: BoundaryCondition
