+++ src/mesh/BoundaryInfo.jl
@@ -1,20 +1,20 @@
 type ElemSidePair
     element::Element
-    side::Int32
+    side::Int64
 end
 
 type BoundaryInfo
     "Unique IDs for sidesets"
-    sidesets::Array{Int32}
+    sidesets::Array{Int64}
 
     "Unique IDs for nodesets"
-    nodesets::Array{Int32}
+    nodesets::Array{Int64}
 
     "Map a sideset ID into an array of element/sides"
-    side_list::Dict{Int32, Array{ElemSidePair}}
+    side_list::Dict{Int64, Array{ElemSidePair}}
 
     "Map a nodeset ID to Nodes"
-    node_list::Dict{Int32, Array{Node}}
+    node_list::Dict{Int64, Array{Node}}
 
-    BoundaryInfo() = new([], [], Dict{Int32, Array{ElemSidePair}}(), Dict{Int32, Array{Node}}())
+    BoundaryInfo() = new([], [], Dict{Int64, Array{ElemSidePair}}(), Dict{Int64, Array{Node}}())
 end
